Switch for charge and current density in Elliptic
In some cases we would like to make a superposition of the externally
applied field (through e.g. embedded boundaries and regular boundary
scalar and vector potentials). In this case, we can calculate a static
electromagnetic field using the static solve through the elliptic
action. The switch potential_only for the elliptic action allows the
chatge and current density registers to be initialised and zerod, but
never calculated.